worker/requirements.txt
celery==4.4.7
flower==1.0.0
yt-dlp==2021.10.10
redis==4.1.2
requests==2.27.1
python-dotenv==0.19.2